# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/raffles": {
      "get": {
        "operationId": "RafflesController_list",
        "summary": "List raffles with optional filters and pagination",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by raffle status",
            "schema": {
              "maxLength": 50,
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Filter by raffle category",
            "schema": {
              "maxLength": 100,
              "type": "string"
            }
          },
          {
            "name": "creator",
            "required": false,
            "in": "query",
            "description": "Filter by raffle creator — must be a valid Stellar public key",
            "schema": {
              "pattern": "^G[A-Z2-7]{55}$",
              "type": "string"
            }
          },
          {
            "name": "asset",
            "required": false,
            "in": "query",
            "description": "Filter by asset code (1–12 alphanumeric characters)",
            "schema": {
              "minLength": 1,
              "maxLength": 12,
              "pattern": "^[A-Za-z0-9]+$",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of records to return",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of records to skip",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raffles list retrieved successfully"
          }
        },
        "tags": [
          "Raffles"
        ]
      }
    },
...
    "/raffles/{id}": {
      "get": {
        "operationId": "RafflesController_getById",
        "summary": "Get raffle detail by ID",
        "description": "Returns merged contract and off-chain details for active, ended, or cancelled raffles (200 OK). Returns 404 Not Found for unknown IDs and 410 Gone for soft-deleted or permanently removed raffles.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Internal raffle ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raffle details retrieved successfully (active, ended, or cancelled)"
          },
          "404": {
            "description": "Raffle not found (unknown ID)"
          },
          "410": {
            "description": "Raffle has been soft-deleted or permanently removed"
          }
        },
        "tags": [
          "Raffles"
        ]
      }
    },
    "/raffles/{id}/participants": {
      "get": {
        "operationId": "RafflesController_getParticipants",
        "summary": "List participants (ticket holders) for a raffle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Internal raffle ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max 100, default 20",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination, default 0",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Participants list retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParticipantListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Raffles"
        ]
      }
    },
...
    "/og/raffles/{id}": {
      "get": {
        "operationId": "OgRenderController_renderRaffleOg",
        "summary": "Pre-render OG meta tags for a raffle (for social media crawlers)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Raffle ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OG Pre-Render"
        ]
      }
    },
...
    "/webhooks": {
      "post": {
        "operationId": "WebhooksController_createWebhook",
        "summary": "Create a new webhook subscription",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "WebhooksController_getWebhooks",
        "summary": "List webhooks for the current user",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/webhooks/{id}": {
      "get": {
        "operationId": "WebhooksController_getWebhook",
        "summary": "Get a specific webhook by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Webhook UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "operationId": "WebhooksController_updateWebhook",
        "summary": "Update a webhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Webhook UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "WebhooksController_deleteWebhook",
        "summary": "Delete a webhook",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Webhook UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Webhooks"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Tikka API",
    "description": "The Tikka API description",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ParticipantDto": {
        "type": "object",
        "properties": {
          "address": {
            "type": "string",
            "description": "Stellar address of the ticket holder",
            "example": "GABCDEFGHIJKLMNOPQRSTUVWXYZ234567"
          },
          "tickets_count": {
            "type": "number",
            "description": "Number of tickets purchased by this address for this raffle",
            "example": 5,
            "minimum": 1
          },
          "purchased_at": {
            "type": "number",
            "description": "Unix timestamp (in seconds) when the first ticket was purchased",
            "example": 1234567890
          }
        },
        "required": [
          "address",
          "tickets_count",
          "purchased_at"
        ]
      },
      "ParticipantListResponseDto": {
        "type": "object",
        "properties": {
          "participants": {
            "description": "List of participants (ticket holders) for the raffle",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ParticipantDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of unique participants",
            "example": 42
          },
          "limit": {
            "type": "number",
            "description": "Number of records returned in this page",
            "example": 10
          },
          "offset": {
            "type": "number",
            "description": "Number of records skipped",
            "example": 0
          }
        },
        "required": [
          "participants",
          "total",
          "limit",
          "offset"
        ]
      },
...
    }
  }
}